feat(amount): update maximum amount value

+++ src/main/java/it/pagopa/ecommerce/payment/methods/controller/v1/PaymentMethodsController.java
@@ -143,7 +143,7 @@ public class PaymentMethodsController implements PaymentMethodsApi {
                                                                                 BigDecimal amount,
                                                                                 ServerWebExchange exchange
     ) {
-        return paymentMethodService.retrievePaymentMethods(amount != null ? amount.intValue() : null, xClientId)
+        return paymentMethodService.retrievePaymentMethods(amount != null ? amount.longValue() : null, xClientId)
                 .map(PaymentMethodsController::paymentMethodToDto)
                 .collectList()
                 .map(
